refactor test coverage as tests.tools
