+++ tests/test_scraper.py
@@ -3,26 +3,10 @@ Test module for the LiveTrailScraper class
 '''
 import unittest
 import pytest
-import inspect
 from scraper.scraper import LiveTrailScraper
 import pandas as pd
+from tools import get_untested_functions
 
-# Define a function to get all functions defined in a module
-def get_functions(module):
-    return inspect.getmembers(module, inspect.isfunction)
-
-# Define a function to get all methods defined in a test class
-def get_test_methods(test_class) -> list[str]:
-    return [method[5:] for method in dir(test_class) if callable(getattr(test_class, method)) and method.startswith("test")]
-
-# Define a function to get unused functions
-def get_unused_functions(module, test_class) -> set[str]:
-    module_functions = set(name for name, _ in get_functions(module))
-    test_methods = set(get_test_methods(test_class))
-    untested_methods = module_functions - test_methods
-    # remove internal class methods (they are tested indirectlyu via the exposed methods)
-    exceptions_test_methods = set(list(filter(lambda x: not x.startswith('_'), untested_methods)))
-    return exceptions_test_methods
 
 class TestLiveTrailScraper(unittest.TestCase):
     # Test case for _checkEventYear method
@@ -273,6 +257,6 @@ class TestLiveTrailScraper(unittest.TestCase):
         assert race_details == scr_race_details, "get_races_physical_details test failed"
 
     def test_implemented_tests(self):
-        unused_functions = get_unused_functions(LiveTrailScraper, TestLiveTrailScraper)
+        unused_functions = get_untested_functions(LiveTrailScraper, TestLiveTrailScraper)
         print(unused_functions)
         assert len(unused_functions) == 0, "LiveTrailScraper is not tested enough. pytest -s for details."
